ddl: remove useless test and adjust test position
